Use `FromReflect` for accessing meta data
